fix: sonar issue fixes and address comments
