+++ projects/flow/src/lib/flow-interface.ts
@@ -1,5 +1,3 @@
-import { FlowComponent } from './flow.component';
-
 export interface ChildInfo {
   position: FlowOptions;
   dots?: DOMRect[];
@@ -25,12 +23,6 @@ export interface DotOptions extends FlowOptions {
   dotIndex: number;
 }
 
-export class FlowConfig {
-  Arrows = true;
-  ArrowSize = 20;
-  Plugins: { [x: string]: FlowPlugin } = {};
-}
-
 export type FlowDirection = 'horizontal' | 'vertical';
 
 export type ArrowPathFn = (
@@ -39,10 +31,3 @@ export type ArrowPathFn = (
   arrowSize: number,
   strokeWidth: number
 ) => string;
-
-export interface FlowPlugin {
-  onInit?(data: FlowComponent): void;
-  afterInit?(data: FlowComponent): void;
-  beforeUpdate?(data: FlowComponent): void;
-  afterUpdate?(data: FlowComponent): void;
-}
