+++ projects/flow/src/lib/plugins/connections.spec.ts
@@ -15,7 +15,8 @@ describe('Connections', () => {
     ];
 
     check(list, [1, 3]);
-    check(list.reverse(), [3, 1]);
+    list.reverse();
+    check(list, [3, 1]);
 
     list = [
       t({ x: 300, y: -150, id: '2', deps: ['1'] }),
@@ -50,30 +51,19 @@ describe('Connections', () => {
     ];
 
     check(list, 0, '2', [1, 3]);
-    // connections = new Connections(list);
-    // let actual = connections.getClosestDotsSimplified(list[0], '2');
-    // expect(actual).toEqual([1, 3]);
     check(list, 1, '1', [3, 1]);
-    // actual = connections.getClosestDotsSimplified(list[1], '1');
-    // expect(actual).toEqual([3, 1]);
 
     list = [
       t({ x: 40, y: 40, id: '1', deps: [] }),
       t({ x: 173.203125, y: -33, id: '2', deps: ['1'] }),
     ];
     check(list, 0, '2', [1, 3]);
-    // connections = new Connections(list);
-    // actual = connections.getClosestDotsSimplified(list[0], '2');
-    // expect(actual).toEqual([1, 3]);
 
     list = [
       t({ x: 40, y: 40, id: '1', deps: [] }),
       t({ x: 142.203125, y: -33, id: '2', deps: ['1'] }),
     ];
     check(list, 0, '2', [1, 3]);
-
-    // actual = connections.getClosestDotsSimplified(list[0], '2');
-    // expect(actual).toEqual([1, 3]);
 
     function check(
       list: ChildInfo[],
