+++ projects/flow/src/public-api.ts
@@ -5,14 +5,10 @@
 export * from './lib/flow.service';
 export * from './lib/flow.component';
 export * from './lib/flow-child.component';
-export {
-  ArrowPathFn,
-  FlowOptions,
-  FlowDirection,
-  FlowConfig,
-} from './lib/flow-interface';
+export { ArrowPathFn, FlowOptions, FlowDirection } from './lib/flow-interface';
 export * from './lib/svg';
 export { FitToWindow } from './lib/plugins/fit-to-window';
 export { ScrollIntoView } from './lib/plugins/scroll-into-view';
 export { Arrangements } from './lib/plugins/arrangements';
 export { Connections } from './lib/plugins/connections';
+export { FlowConfig, FlowPlugin } from './lib/plugins/plugin';
